CI実装

+++ backend/app/crud/content_crud.py
@@ -2,26 +2,27 @@ from boto3.dynamodb.conditions import Key, Attr
 from app.schemas.content import RegisterContentData, ContentData, watchlistData
 from typing import List, Any
 
+
 def add_content(content: RegisterContentData, table: Any):
     item = content.dict()
     table.put_item(Item=item)
 
+
 def update_content(content: RegisterContentData, table: Any):
     response = table.update_item(
-        Key = {
-            'contentId': content.contentId,
-            'userId': content.userId
-        },
-        UpdateExpression="SET #ty = :ty, #ti = :ti, #d = :d, #y = :y, #n = :n, #l = :l REMOVE #s",
-        ConditionExpression="attribute_exists(userId) AND attribute_exists(contentId)",
-        ExpressionAttributeNames = {
+        Key={"contentId": content.contentId, "userId": content.userId},
+        UpdateExpression="SET #ty = :ty, #ti = :ti, #d = :d, #y = :y, #n = :n,\
+            #l = :l REMOVE #s",
+        ConditionExpression="attribute_exists(userId) AND\
+            attribute_exists(contentId)",
+        ExpressionAttributeNames={
             "#ty": "type",
             "#ti": "title",
             "#d": "date",
             "#y": "year",
             "#n": "notes",
             "#l": "link",
-            "#s": "status"
+            "#s": "status",
         },
         ExpressionAttributeValues={
             ":ty": content.type,
@@ -29,61 +30,60 @@ def update_content(content: RegisterContentData, table: Any):
             ":d": content.date,
             ":y": content.year,
             ":n": content.notes,
-            ":l": content.link
+            ":l": content.link,
         },
-        ReturnValues="ALL_NEW"  # 更新後の完全なアイテムを返却
+        ReturnValues="ALL_NEW",  # 更新後の完全なアイテムを返却
     )
-    return response.get('Attributes')
+    return response.get("Attributes")
+
 
 def update_best(contents: List[ContentData], table: Any):
     for content in contents:
         table.update_item(
-            Key= {
-                'contentId': content.contentId,
-                'userId': content.userId
-            },
+            Key={"contentId": content.contentId, "userId": content.userId},
             UpdateExpression="SET #rnk = :r",
             ExpressionAttributeNames={
                 "#rnk": "rank"  # rank をエイリアス #rnk として定義
             },
-            ExpressionAttributeValues={
-                ":r": content.rank
-            }
+            ExpressionAttributeValues={":r": content.rank},
         )
+
 
 def delete_best(contents: list[ContentData], table: Any):
     for content in contents:
         table.update_item(
-            Key= {
-                'contentId': content["contentId"],
-                'userId': content["userId"]
+            Key={
+                "contentId": content["contentId"],
+                "userId": content["userId"],
             },
             UpdateExpression="REMOVE #rnk",  # エイリアスを使用
             ExpressionAttributeNames={
                 "#rnk": "rank"  # rank をエイリアス #rnk として定義
-            }
+            },
         )
+
 
 def get_years(userId: str, table: Any):
     response = table.query(
         IndexName="YearIndex",
         KeyConditionExpression=Key("userId").eq(userId),
         ProjectionExpression="#yr",
-        ExpressionAttributeNames={
-            "#yr": "year"  # year を予約語として回避
-        }
+        ExpressionAttributeNames={"#yr": "year"},  # year を予約語として回避
     )
 
     # `year`のみのリストを取得し、重複を排除
     years = list({item["year"] for item in response.get("Items", [])})
     return years
 
+
 def get_year_contents(user_id: str, table: Any, year: int) -> list[dict]:
     response = table.query(
         IndexName="YearIndex",
-        KeyConditionExpression=Key("userId").eq(user_id) & Key("year").eq(year)
+        KeyConditionExpression=Key("userId").eq(user_id)
+        & Key("year").eq(year),
     )
     return response.get("Items", [])
+
 
 # TODO:不要そうであれば後ほど削除
 # def get_year(user_id: str, year: int) -> list[dict]:
@@ -93,17 +93,20 @@ def get_year_contents(user_id: str, table: Any, year: int) -> list[dict]:
 #     )
 #     return response.get("Items", [])
 
+
 def get_year_best(user_id: str, year: int, table: Any) -> list[dict]:
     response = table.query(
         IndexName="RankIndex",
         KeyConditionExpression=Key("userId").eq(user_id),
-        FilterExpression =Attr("year").eq(year)
+        FilterExpression=Attr("year").eq(year),
     )
     return response.get("Items", [])
+
 
 def add_watchlist(content: watchlistData, table: Any):
     item = content.dict()
     table.put_item(Item=item)
+
 
 def get_watchlist_contents(user_id: str, table: Any) -> list[dict]:
     """
@@ -114,9 +117,11 @@ def get_watchlist_contents(user_id: str, table: Any) -> list[dict]:
     :return: ウォッチリストのコンテンツ（リスト形式）
     """
     response = table.scan(
-        FilterExpression=Attr("userId").eq(user_id) & Attr("status").eq("to_watch")
+        FilterExpression=Attr("userId").eq(user_id)
+        & Attr("status").eq("to_watch")
     )
     return response.get("Items", [])
+
 
 def delete_watchlist_contents(user_id: str, table: Any, content_id: str):
     """
@@ -127,9 +132,4 @@ def delete_watchlist_contents(user_id: str, table: Any, content_id: str):
     :param content_id: 削除対象のコンテンツコンテンツID
     :return: なし
     """
-    table.delete_item(
-        Key={
-            "contentId": content_id,
-            "userId": user_id
-        }
-    )
+    table.delete_item(Key={"contentId": content_id, "userId": user_id})
